Bug #3926: Tomcat servlet container did not provide a class
It seems that Jetty provides "com.fasterxml.jackson.databind.JsonNode"
in any way but Tomcat does not.

So we have to include the package containing this class in our Maven
configuration.

We seriously have to check if ARSnova runs on our main target platform,
not only our private developer platform! My be our developer platform
should be the main target platform?