Fixed wrong handling of nonexisting feedback entries
If some kinds of feedback values do not exist do not throw the exception
NotFoundException. This only indicates, that these values should be '0'
which is the default value.

NotFoundException should only be thrown if the requested session of
feedback does not exist.
Every session has a feedback - sometimes with only 0 values.