Major changes is Spring application configuration
With this changeset the spring servlet configuration is separeted from the
application configuration. This will result in a main context and a servlet
context, holding only spring mvc information.

This change has a very nice side effekt: Spring AOP beans will be loaded
only once in application context, not a second time in mvc servlet context.