Add UserSessionAspect aspect
This sets current user and ARSnova session information in a session
scoped service bean. If the user connects with a websocket, the
websocket session ID will be set in addition to the user and session
object.

The aspects methods pointcut after returning from joinSession() method
in SessionService. So the UserSessionService will be set with all needed
data after a user successfully joins a session.

With this service available, it can be used to replace some usage of
Array/Map based application wide data structures that cause some trouble
when stopping or redeploying the application.