Added property _conflicts to avoid NoSuchMethodException
CouchDB tries to set _conflicts propertiy but did not find a setter and
throws NoSuchMethodException. This patch will avoid exception throwing.

The reason of exception throwing can be found can be found in CouchDBs
availibility to store different and concurrent versions of one single
document. This error occurs if two different versions of one session
document are stored within CouchDB.

Further we have to take care, that there is only one version of a
session, not multiple versions.