Deprecate getSession, use getSessionFromKeyword instead
getSession throws an HTTP exception, which is not
good style. Additionally, as it calls
getSessionFromKeyword internally -- which has a caching
annotation -- this caching mechanism will not work
because it only works if the method is called
from another object. This is due to the way Spring
is using proxies to intercept the calls.