Fix #15667: Skip over documents with errors
If deleted documents remain in the sort order,
this will crash the application. We simply skip
over any documents we could not load.

With this fix, deleted documents may remain
in the sort order without a problem. Once the
order is updated by the teacher, any deleted
documents will be removed.