Initial idea of a count service
This new service should keep the number of questions
up to date without hitting the CouchDB every time.

Currently, it only counts the total number of questions
according to their variant, but this could be extended
to the number of answers and even to the number of
unanswered questions per user.