To prevent address conflicts for Socket.IO use upcounting port offset
In some cases the Socket.IO servers used in test cases are not closed
when starting a new test. In this cases the tests will fail with
an address conflict because an other Socket.IO server is still running.

Using a new port, starting with previous defined port 1234 up to 1243 will
assign every new created Socket.IO server an other address/port.

An address conflict will still occure if an other service is running on
ports from 1234 to 1243, so it is generally a bad idea to have tests in the
hope that no side effects will occure from outside of application scope.