Add additional profile to exclude Controller from other tests
This solves the IllegalStateExceptions occurring for most of the other
test cases.

However, we now have the old problem that different Repositories are
injected into objects. In my case, the testCaching case of the
DefaultEntityServiceImplTest stops working, because it gets a different
RoomRepository than the ApplicationPermissionEvaluator.

This behavior is caused by the introduction of the additional profile.