Use DefaultEntityServiceImpl for most UserProfile operations
Some account operations needs to be run without a fully authenticated
user so the service directly accessed the repository. A run-as role is
now set to allow the use of the default implementation.

With this change, events are now published for CRUD operations on
UserProfiles and issues with caching should be fixed.