Fix testing configuration
A configuration has been added for the Maven Surefire plugin to
* explicitly set its version, and
* let the phase fail if no tests were run.

This change was needed because Maven by default used an outdated version
of the plugin which was not compatible with JUnit 5's Jupiter engine, so
those tests were skipped. The CI job failed because no output was
generated.

The `unit_test` CI job has been reenabled.