Add skipMissing parameter for /?ids= endpoint
The introduced `skipMissing` query parameter allows to customize the
behavior for missing (not found or no read permission) entities. The
default value `true` implements the previous behavior of skipping the
missing values. When set to `false`, `null` values are used to fill the
holes in the result list, so the `List`'s size and order match the list
of requested IDs.