Handle missing entities in downstream
When the user queries a room via room summary view that doesn't exist or the user doesn't
have access to, the gateway will return null in that spot.

When the user for some reason has a room access entry for a removed room, it'll get
filtered out in the membership view.