Improve featureset handling and fix math in Markdown
A minimum featureset (new lines, entites) have been added which is used
when Markdown rendering is disabled. Also, math features are added to
the featureset if LaTeX is enabled, so it renders once again in
combination with Markdown. Furthermore, in LaTeX-only mode, LaTeX is now
rendered inside text delimited by dollar signs.