Publish a new BulkChangeEvent
The `BulkChangeEvent` is published after multiple entities have been
patched or deleted. It can (and should) also be used after bulk creation
but there currently is no generalized bulk creation method for entities,
so publishing needs to be handled explicitly by the respective entity
service. This is the case for the AnswerService's scheduled bulk
creation.