Remove obsolete code for eventual consistency for room access and listen to generic room events
This introduces new routes to let the gateway create room access, instead of listening to events from the backend.
It removes the need for the sync tracker for most functionality excluding the explicit syncing of a room.
It also teaches the auth service to listen to generic room events (when a room is created or deleted) and tries to add/remove access accordingly.