Add room auth filter to update room access at auth service
Introduce a new filter that listens to changes in room access.
Whenever a room is created, deleted, transfered, getting or losing a new moderator,
a request to the auth service is chained after the core has handled it.
This lets the core handle business logic first and only if that part of the requests succeeds,
the auth service is informed about new access.

This allows for completely removing room access handling in the core.