Improve handling of the removed subject for old contents
Subeject is no longer used by the current web client. But still, the
subject is set for previously created contents.

This change prepends the subject to the body (if it is set) and extracts
the subject from the body for the legacy API. Furthermore, the subject
is no longer required to be set.