Track metrics for creation and deletion events
Event metrics are tracked for the following entity types:
* UserProfile
* Room
* Content
* Answer

Additional metrics for `UserProfile` events are tracked which are tagged
with the auth provider type (`account`, `guest`, `sso`).

A metric for the count of survey votes has also been implemented but
doesn't currently work because survey events (`NewFeedbackEvent`) are no
longer published.