Remove invalid allowedOrigins setting for /ws endpoint
Setting a wildcard for the CORS origin header is invalid when sending of
credentials is allowed. Spring now throws an InvalidArgumentException if
this combination is used:

    When allowCredentials is true, allowedOrigins cannot contain the
    special value "*"since that cannot be set on the
    "Access-Control-Allow-Origin" response header. To allow credentials
    to a set of origins, list them explicitly or consider using
    "allowedOriginPatterns" instead.

We should not need CORS because the same origin is used for requests to
the /ws endpoint. If it is needed in the future, allowed origins should
be set explicitly.