Record cache metrics
The following metrics are now tracked for Caffeine trough Micrometer:
* size
* puts
* gets (hits and misses)
* evictions

These metrics are recorded separately for each cache. The caches are now
statically created which is a requirement of Actuator so it can
automatically setup registries with Micrometer. Using a cache which has
not been defined at startup will now lead to an exception.