Add new properties for publishing Contents to ContentGroup model
New properties have been added to set the range of published Contents of
a ContentGroup instead of setting the state for each Content
individually.

A Content is now published if its ContentGroup is `published` and for
its index `i` inside `contentIds` applies:

    firstPublishedIndex <= i <= lastPublishedIndex

or

    firstPublishedIndex <= i && lastPublishedIndex == -1

Furthermore, the contentIds are now stored as a List (instead of a
LinkedHashSet) because to entry index inside the Collection gained
importance. The uniqueness is now checked by domain logic. This is an
implementation detail and does not affect the API and persisted data.
The use of a List makes checking if a Content's index is in the
published range much easier.