Add a migration to move state from Contents to ContentGroup
This migration moves some state information previously stored for
individual Contents to ContentGroups. The new publishing state and range
for ContentGroups is determined by first checking if all or none of the
individual Contents are published (`state.visible`). In this case, the
published is set accordingly and `-1` is set for `firstPublishedIndex`
and `lastPublishedIndex`. Otherwise, published is set to true and the
range is determined by taking the first published Content and the last
published Content in a row. Later Contents, which are not published
consecutively, will be excluded from the range.

Furthermore, this migration adds additional properties which have been
introduced to the models and removes obsolete ones.