Use caching for multi-entity requests
When multiple entities are retrieved by ID, the cache is now checked for
each ID. Only the entities which are not yet cached are retrieved from
the database and then put into the cache.

This does not cache results of database view queries. But those queries
are primarily used to fetch a list of IDs of entities which are then
retrieved in a separate step. This in turn can now use the above
mentioned caching behavior.

Closes: #97