Always retain entity ordering as defined by order of passed IDs
The `DefaultEntityServiceImpl` now retains the order for entities
returned by `get(Iterable<String> ids)`.

!363 fixed ordering for entities only on the controller level under the
wrong assumption that the order would not matter on the service level.
The order is important for the content IDs of content groups. The
changes have been reverted because reordering on the controller level is
no longer necessary.

Fixes: #102
Regression-MR: !361
Regression-Commit: a4459240ed9ef3df0a74c62a3bf674e97ee055a0
Related: !363