Track metrics for running sessions in rooms
The service considers that a session is running in a room if a certain
threshold of users is reached. The session is considered as ended when a
drop of users is detected. The maximum user count and the duration are
tracked once per session to create a distribution summary.

Related: particify/dev/foss/arsnova-ws-gateway#12