Fix unintentionally skipped documents during migrations
The next batch of entities is now queried before the current ones are
updated in the database.

This is a workaround for a CouchDB issue with bookmarks. When the last
document of a page is changed in a way that is would no longer match the
query's selector, the next page is skipping the first expected document.

Fixes: #120
See: https://github.com/apache/couchdb/issues/3657