Enable merging for nested POJOs of entity models
Previously, if properties of a nested POJO were changed through a patch
operation, other properties of the nested object needed to be explicitly
set. Otherwise, they would be overridden with the default value.

Merging has now been explicitly been enabled for nested properties where
it makes sense, i.e. properties which are accessible through the API for
patching. It has not been enabled for extensions Maps because this would
have the unwanted effect of also enabling merging for nested Lists.

Closes: #111