Add WS events for changes of room stats
New events are published for room stats. Specific changes cannot be
detected for non-entities, so the whole object is sent as payload.

Two topics are used, one for publicly visible and one for moderator
roles. These topics are intended to be used for other non-entity changes
in the future.

STOMP topics:
* `<Room ID>.changes.stream`
* `<Room ID>.moderator.changes.stream`

Event structure:
The full `RoomStatistics` object is sent without wrapper. The STOMP
header `__TypeId__` can be used on the client side to filter events by
type.

Related: particify/dev/common-concerns#14