Add service and endpoints for entity duplication
Duplication has been implemented for rooms and contents. Room
duplication cascades, i.e. entities related to the room are also
duplicated and assigned to the new room. Content duplication adds a copy
of a content and adds it to the same or another content group. Answers
for those contents are not duplicated.

Endpoints:
* `POST /room/<Room ID>/duplicate`
* `POST /content/<Content ID>/duplicate`
  `{"contentGroupId": "<Target Content Group ID>"}`

Closes: #123