Adjust CSV import/export to SCALE format changes
The import and export needed adjustments for SCALE format contents
because of the recent changes to the format. Instead of storing the
individual options, the template and option count are written to the
options field (e.g. `AGREEMENT (5)`).

When previously exported `SCALE` contents are imported, the format is
changed to `CHOICE`.

Closes: #126
Related: particify/dev/common-concerns#10, !394