Remove legacy security annotations
With the refactoring of !345 authorization handling has been extracted
to a separate layer, so controllers could use the new layer while
internal operations could bypass the security checks.

The security annotations were not removed in the `EntityService`
interface. As a consequence, security checks were still performed for
basic CRUD operations even though an unsecured service was used.

With the removal of those annotations from `EntityService`, it is no
longer necessary to switch roles for internal operations (`RUN_AS`) and
entity service methods can be used by operations which are not triggered
by a user (e.g. `@Scheduled`).

Related: !345