Add new, modular configuration for Docker Compose
The new configuration setup uses multiple compose configurations and env
files for enhanced flexibility and to allow reuse by the SDK. Docker
Compose is no longer used directly. Instead, a wrapper script is
provided (`particifyctl`) which merges the building blocks of the
configuration and passes appropriate arguments to Docker Compose.

Separate files are now used for local configurations: `local.env` and
`local.compose.yaml`. This way, there are no longer conflicts when the
configuration was changed locally and in the upstream repository.