Simplify content persistence logic
Some content logic was still built around the no longer existing variant
property which was replaced with the introduction of content groups.
This logic has been simplified or removed where it was no longer needed.

As a side-effect, this refactoring also enables caching for contents
which are queried by room ID. This shouldn't have a huge impact because
contents are retrieved directly through their IDs from the content group
in most cases.