Fix DuplicateKeyException in DefaultEntityServiceImpl
An exception was thrown when a list entity IDs with duplicates was
resolved from cache in
`DefaultEntityServiceImpl.get(Iterable<String> ids)`. Duplicate IDs are
now filtered.

This previously could lead to errors in user answering stats calculation
when a content group contained duplicate content IDs. This is not
supported but the creation of such groups is currently not prevented on
the backend side.

Related: !585