Add auto delete functionality for inactive accounts
The time of last account activity is determined by the last access of
any room. The core requests a list of IDs of inactive users and deletes
the corresponding accounts on a schedule.

Auto deletion is disabled if a user inactivity period of 0 days is set.
This is the default. As a precaution, a limit for the number of affected
accounts can be set.

Settings:
* system.auto-deletion-thresholds.user-inactivity-period (default: 0D)
* system.auto-deletion-thresholds.user-inactivity-limit (default: 50)

Closes: #185