Add /healthz endpoint
This adds a /healthz endpoint which does not require JWT authentication
but is only accessible from site-local networks. This provides an easy
way to perform basic monitoring.

For now, this just passes the health info from the core. This can be
extended in the future, to provide aggregated health info from all
services.

Closes: #204