Add setting to override local hostname for SMTP
Java automatically uses the local hostname for sending mails and uses it
for EHLO and the `Message-ID` header's scope. When running in a
containerized environment, the hostname of the container might be used
internally only and will not be resolvable by target mail servers. This
should not cause problems with mail sending but can make tracing in case
of problems harder.

For Docker Compose, the hostname of the webserver is reused as localhost
for SMTP.

Closes: #209