Add support for JWT generation to authz
A new `/jwt` endpoint has been added for internal use, which generates
a JWT with room roles for a user.

The endpoint expects a JWT and an URI to check room authorization. If
the user is authorized, a new JWT with the role is included in the
Authorization header of the response. Otherwise, a 403 response is sent.

Required request headers:
* Authorization
* X-Forwarded-Uri