Add CSV import/export handling for NumericContent
Four fields specific to the format need to be handled: min, max, correct
and tolerance. To avoid adding format specific columns to the CSV, the
options and correctOptions columns are reused. Min and max are stored in
options, separated by a minus sign. Correct and tolerance are stored in
correctOptions, separated by a plus-minus sign.

Closes: #230