Simplify content group model for content publishing
This introduces a new publishing mode with a single index which replace
the published state and publishing range.

Value for `publishingMode`:
  * NONE: Publish no contents / do not publish content group at all
  * ALL: Publish all contents
  * SINGLE: Publish a single content
  * UP_TO: Publish contents up to index

The model for statistics has been adjusted to reflect this change and
now includes the count of content groups by mode instead of the count of
published ones and those using publishing range.

Closes: #236, #237