Add leaderboard for content groups
The new leaderboard endpoint returns a list of all users which have
participated in any content of the group. Each entry includes the alias
object for the user and the total score. If a contentId is set, it also
includes the result for the content with points, duration in ms and a
flag for correctness.

Endpoint:

    /room/{roomId}/contentgroup/{groupId}/leaderboard[?contentId={contentId}]