Add new displayId property to UserProfile
The loginId has been used as a public user ID which is shown on the UI
and needs to be known to other users so room permissions can be
assigned. When an SSO is used, the loginId is set to an SSO-provided
attribute to map the user. In some cases, the attribute used for the
mapping is not appropiate to be used as a public user ID.

The displayId has been introduced to provide more flexibility and allow
another attribute to be used as a public user ID. The attribute for the
display ID can be set via configuration. Currently, loginId (default)
and mail are supported for use as displayId.

On the frontend-side, the displayId should now be used everywhere where
previously the loginId was used except for the login process itself.

Closes: #248