Cache leaderboard score data
Content and total scores for the leaderboard are now cached. Content
score cache entries are evicted when they become outdated. For the total
scores, the whole cache has to be cleared when the answer queue is
flushed (i.e. every 5 seconds) because there is currently no way to
determine the cache key without significant overhead.

Closes: #252