Increase points for correctly answered contents
The points for correct answers have been increased to 500. The points
are multiplied by 2 if a duration is set for the content. This way,
differences in answering times are better reflected in the points.
Contents with a duration have a maximum of 1000 points but because the
remaining duration is factored less points would be assigned. A content
that is answered when half of the duration is remaining results in the
same amount of points that a question without a duration would give.

Related: particify/dev/common-concerns#67