Store result in answer and add stats endpoint
A new /stats endpoint has been added for content groups which allows to
retrieve result stats for all contents in a group with a single request.

For performance reasons the result is now stored in answers, so it does
not need to be evaluated every time. This commit does not include a
migration for existing answers, so stats only are available for contents
with all new answers.

Closes: #255