Add LIVE and remove SINGLE mode for content groups
A new publishing mode `LIVE` for live use cases has been added.
For non-live use cases, only publishing of all and and a range of
contents is relevant. For live use cases, Publishing of a range and
single publishing is relevant.

Furthermore, the `SINGLE` mode has been removed. With the existence of
focus mode for rooms, it didn't not provide much value and added
complexity.

Closes: #258
Related: particify/dev/common-concerns#69