Add external login mapping and SAML support
Users can now be linked to ExternalLogins. The models support a single
user having a local login and multiple external logins. But for now,
there is no logic to link additional external logins to an existing
user. A new user is created on first login.

Futhermore, a SAML implementation as first provider for external logins
has been added. This implementation has the following features:

* multiple SAML IdPs can be added
* IdPs are configured through metadata URLs
* mapping of SAML attributes to user properties
* request signing

Closes: #322
Related: #325