Add Keycloak and config script for dev environment
A Keycloak service is now started as part of the development Docker
Compose environment. A realm template configuration and a script for
certificate generation is provided, so it works out of the box for
authentication using the SAML protocol.

The core4/generate-credentials.sh script need to be run once to generate
the private and public key for the service provider and include the
public key in the generated realm config.

Closes: #326