Add read-only mode
When enabled, this mode disables the creation and modification of most
data. It does however not prevent the deletion of data. Such requests
will fail with status 403 in this case. Furthermore, it overrides the
room-specific read-only setting for the Q&A.

UserProfiles are excluded from read-only mode handling because this
could lead to login issues.

Closes: #332