Fix handling of duplicate room user alias keys
In rare cases it can happen that multiple aliases are created for a user
in a room, which breaks the leaderboard. This fix does not prevent the
creation of multiple aliases because CouchDB has no concept of
uniqueness. So the application would need keep track of aliases
independently from the database. Instead, additional aliases are now
ignored when creating the mapping for users.

Closes: #342