Add implementation for granting role by invitation
The `grantRoomRoleByInvitation` mutation invites a user, which does not
have an account yet. A new user is created, an invitation mail is sent
and the role is granted to the unverified account.

A separate mail template is used for invitations. In contrast to the
usual verification, it does not include the code directly. A code could
be confusing for a user who has not requested the verification mail
themself.

While an unverified account can now be granted a role, this is only
possible through the new invitation process. Pending invites are
included when querying managing users. The client can use the user's
verified flag to differentiate.

Closes: #352