Improve response and error handling
Custom exceptions and resolver adapters are now used for GraphQL. A
classification extension field is set for the GraphQL error based on the
thrown exception. These classifications allow the client to
differentiate between errors without having to check for the exact
message.

Resonse types have been adjusted so they include the changes instead of
a boolean flag for the success status.

Closes: #360